Zadanie9/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
google-generativeai==0.3.2
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6